Update SVD Module with static Options parameter
Adds Options template parameter to the SVD module to allow
requesting unitaries statically. This update is applied to both
BDCSVD and JacobiSVD. This deprecates the SVD constructor that
takes Unitary options as an argument. It also removes the overload
of compute that accepts options. Finally, this includes a
Refactoring of the SVD tests and internal uses of SVD.