Port EIGEN_OPTIMIZATION_BARRIER to soft float arm
On arm, floating point hardware is optional. Constraint "w" does not
compile if the target platform does not have either VFP or NEON, for
example arm1136j-s. This change removes the "w" option when macro
`__ARM_FP` is not defined.

To test this, add "+nofp" on the -march CFLAG or select a CPU without
FP using the -mcpu CFLAG.