Improve clang warning suppressions by checking if warning is supported
The previous technique of looking at the clang version is fragile; instead use the `__has_warning` mechanism, which clang has supported for a very long time.

This fixes a warning with AppleClang 10.0 which reported:

```
warning: unknown warning group '-Wimplicit-int-float-conversion', ignored [-Wunknown-warning-option]
    #pragma clang diagnostic ignored "-Wimplicit-int-float-conversion"
                                     ^
```

Also removed the suppression of -Wabsolute-value since the version check it was wrapped in would evalute to false in all versions of clang for many years now. The check was probably trying to do 'if version >= 3.5', but wasn't.