Re-add `svd::compute(Matrix, options)` method to avoid breaking external projects.
There are too many other projects (open-source, and at Google) that rely
on the existing mechanism.  For the open-source projects, we at least need a
version number increase to check for which API to use.  We also likely
need some transition time to allow projects to adapt.

Adding the method back in seemed trivial.  We may want to reconsider whether
we truly want to deprecate this behavior.  If we do, we can probably
remove it after the next major release.

/cc @rmlarsen1 @arthurfeeney