Fix gcc-5 packetmath_12 bug.
There seems to be a gcc-5 bug that's causing `data1` within
`packetmath_minus_zero_add()` to be filled with
```
(-0, 0), (-0, NaN)
```
when optimizations `-O2` or higher are on.  This is before any packet
operations are called.  Printing the values causes the bug to disappear.  I've
double-checked we're not running into any aliasing bugs (the cast from
`std::complex<double>*` to `double*` is legal, and using c++ casts
doesn't fix the issue).  Initializing the memory causes the bug to
disappear.  The test also works with gcc-6 and later.