Fix packetmath compilation error.
Unfortunately we can't pass a pointer to `psqrt<Packet>` as a functor, since such a
function might not exist (e.g. if `HasSqrt` is `false`).  The only way to pass an overloaded function
as a functor is to wrap it in a struct.  Here we create a simple wrapper
around `psqrt`, `prsqrt`.