Update vectorization_logic tests for all platforms.
A few of the tests depend on packet (and sub packet) availability,
implementation of specific packet functions (e.g. add, div), and
unrolling limits and costs, which can vary by platform.  Here we add
the ability to ignore the traversal or unrolling when
it is not informative.

Also, with some half-packets, if the size of the packet is smaller
than 16 bytes (e.g. `Packet2f`), then matrices of certain sizes
no longer vectorize with `EIGEN_UNALIGNED_VECTORIZE=0`, since they
are assumed unaligned.  Adjusted matrix sizes in such cases to
get the tests to pass.

This *should* help tests pass on all platforms.  Tested with
SSE, AVX, AVX512, NEON, AltiVec.