Fix bug when solving wide matrices
There is a unit test for this case, but the unit test has a bug too.
It's accidentally making a TALL matrix, not a WIDE matrix.
The bug in the code is that `solveInactiveSet_` is using a RHS vector
as workspace for RHS-sized vectors and solution-sized vectors.
When the solution has more entries that the RHS, this breaks.