Avoid ambiguous Tensor comparison operators for C++20 compatibility
C++20 automatically tries looking up the reversed form of a comparison
operator, so we end up with with an overload set of
operator==(const TensorBase<Derived>&, const OtherDerived&)
operator==(const OtherDerived&, const TensorBase<Derived>&)
These two are ambiguous if both types are the same. Explicitly spell out
that OtherDerived is a TensorBase to resolve the ambiguity.